Dynamic div size
I want to change the size of the divs dynamically
so I set the size as a variable in the script,
and then change the container's sizes accordingly.

This adds inline style to the HTML tho, which is
really bad and ugly, and yucky, and I don't want it.

Also I should delete setAttribute.